refactor: bestMeetimeList get 하는 hook 리팩토링
